Implemented create command
Implemented the create command to create a new zellij session with a
name, path, and layout. When in an existing zellij session, this command
will use the zellij-switch plugin to create the new session and switch
to it. The plugin doesn't check if the session already exists, so that
check still needs to be done here.